WARNING! This version crashes - but I don't know why and I please for help. Summary follows.
1. Added new (abstract) classes: NormalInteraction (with kn (normal) sitffness) and NormalShearInteraction (additionally, with ks (shear) stiffness). They are used by GlobalStiffnessCounter. 
2. Removed SimpleElasticInteraction, since it it basically NormalInteraction, and changed all code to reflect that.
3. Changed GlobalStiffnessCounter to work with the NormalShearInteraction class, since it provides an abdstract interface and will be useful with more classes than just ElasticContactInteraction.

Now, the problem:

With TriaxialTest in GlobalStiffnessCounter, I get crash that I am not able to handle: ElasticContactInteraction (which is used in that simulation) should be dynamic_cast'd to NormalShearInteraction, which is its direct parent. But the cast fails every time. There is a line putting the class name to cerr and it really is ElasticContactInteraction.

Can someone suggest a solution?

Many thanks!