1 - stress definition in PeriIsoCompressor was wrong for ScGeom (sign mistake due to double-reverse);
2 - applyForceAtContactPoint cannot be used in periodic + ScGeom, since this function needs the wrapped position of each body (in Dem3Dof, wrapped position is an 
attribute, in ScGeom there is no such attribute and we don't want to define it as it costs cpu). The function call is replaced by separate applyForce/applyTorque using a 
"branch" computed using radius, normal, and penetration depth.

3 - Implement homothetic=2 properly. The acceleration of the period (i.e. change in velGrad) is applied on particles as an equivalent impulse. If velGrad is constant, 
there is no effect on particles. If velGrad changes, each particle is accelerated with (delta_velGrad*position). This is like giving initial velocities to particles so 
that they move homothetically.

TODO :
4 - Define relative velocity across period correctly. It should be done adding a velGrad*shift2 term, but shift2 is unknown in Law2... I'll maybe exploit the "shear" 
attribute of ScGeom to precompute a shifted shear disp. increment in Ig2, and use it later in Law2.
(btw : isn't that an interesting option to merge ScGeom and Dem3? Letting shear be the total OR incremental displacement?)